Add: combine and filter tweets
Because the tweets returned by the Twitter Streaming API don't always
match in a way that we would expect, sometimes an additional layer of
filtering is required.

As an example, sometimes we want only those tweets that have a specific
coordinates to be matches. However, the filter endpoint will try to use
the region defined in the `place` field if the `coordinates` field is
not populated, which will cause false positive matches.