Add: Compute census block interactions
We can model movement through a region by computing census block
interactions. To do so, we use the chronological order of geocoded
tweets for each user to:

- Find the census block that contains the tweet.
- For every pair of tweets (edges), count the number of interactions
between the source block and the target block.

Currently the census block spatial polygons are **not** indexed,
therefore the computation takes a significant amount of time. Indexing
or other optimizations should be performed, otherwise the interactions
should be computed only once if possible.